+++ tomcat/src/main/java/com/techcourse/controller/StaticController.java
@@ -1,8 +1,8 @@
 package com.techcourse.controller;
 
 import java.nio.charset.StandardCharsets;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 
 public class StaticController extends AbstractController{
